Split generate.py
